REFACTOR : UserService 의존성 관련 리팩토링 #14
